Adding Navigation, navigation header, search, Add Rides, and Home page.
